+++ src/doom/r_segs.cpp
@@ -258,7 +258,7 @@ R_RenderMaskedSegRange
 	{
 	    if (!fixedcolormap)
 	    {
-		index = spryscale>>(LIGHTSCALESHIFT + crispy->hires);
+               index = (unsigned int)(spryscale>>(LIGHTSCALESHIFT + crispy->hires));
 
 		if (index >=  MAXLIGHTSCALE )
 		    index = MAXLIGHTSCALE-1;
@@ -382,7 +382,7 @@ void R_RenderSegLoop (void)
 	    texturecolumn = rw_offset-FixedMul(finetangent[angle],rw_distance);
 	    texturecolumn = texturecolumn >> FRACBITS;
 	    // calculate lighting
-	    index = rw_scale>>(LIGHTSCALESHIFT + crispy->hires);
+	    index = (unsigned int)(rw_scale>>(LIGHTSCALESHIFT + crispy->hires));
 
 	    if (index >=  MAXLIGHTSCALE )
 		index = MAXLIGHTSCALE-1;
